Rename memory block files
