+++ src/plugins/intel_cpu/src/nodes/gather.cpp
@@ -4,7 +4,7 @@
 
 #include "gather.h"
 
-#include <partitioned_mem_mgr.h>
+#include <partitioned_mem_blk.h>
 
 #include <cstdint>
 #include <openvino/op/constant.hpp>
