+++ src/plugins/intel_cpu/src/nodes/memory.hpp
@@ -7,7 +7,7 @@
 #include "input.h"
 #include "memory_state_base.h"
 #include "ov_optional.hpp"
-#include "proxy_mem_mgr.h"
+#include "proxy_mem_blk.h"
 
 #include <map>
